requirements.txt
beautifulsoup4==4.11.1
camelot-py==0.10.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
et-xmlfile==1.1.0
ghostscript==0.7
idna==3.4
numpy==1.24.1
opencv-python==4.7.0.68
openpyxl==3.0.10
pandas==1.5.3
pdfminer.six==20221105
pycparser==2.21
PyPDF2==2.12.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.9.0
urllib3==1.26.14
XlsxWriter==3.0.7
